refactor(world): split movement.rs into submodules (TD-027)
Behavior-preserving split of parish-world/src/movement.rs (~899 LOC)
into movement/mod.rs (hub re-export), movement/resolve.rs (types +
target resolution + public entry points), movement/transport_apply.rs
(weather/transport adjustment + narration builders), and movement/tests.rs.
Public API unchanged. Part of #1200.
